port wr project to _WIN64

+++ bld/sdk/wr/h/wrmemi.h
@@ -38,7 +38,7 @@
 /****************************************************************************/
 extern void WRMemOpen( void );
 extern void WRMemClose( void );
-extern void *WRWResMemAlloc( unsigned );
+extern void *WRWResMemAlloc( size_t );
 extern void WRWResMemFree( void * );
 
 /* functions to replace those in mem.c in commonui */
